+++ R/addGroupOfUnusedAnimals.R
@@ -7,9 +7,21 @@
 #' of animal Ids.
 #' @param candidates character vector of IDs of the animals available for
 #' use in the group.
+#' @param ped dataframe that is the `Pedigree`. It contains pedigree
+#' information including the IDs listed in \code{candidates}.
+#' @param minAge integer value indicating the minimum age to consider in group
+#' formation. Pairwise kinships involving an animal of this age or younger will
+#'  be ignored. Default is 1 year.
+#' @param harem logical variable when set to \code{TRUE}, the formed groups
+#' have a single male at least \code{minAge} old.
 #'
 #' @export
-addGroupOfUnusedAnimals <- function(savedGroupMembers, candidates) {
+addGroupOfUnusedAnimals <- function(savedGroupMembers, candidates, ped,
+                                    minAge, harem) {
+  if (harem) { # Sires were added to groupMembers
+    candidates <- removePotentialSires(candidates, minAge, ped)
+  }
+
   # Adding a group for the unused animals
   n <- length(savedGroupMembers) + 1
   savedGroupMembers[[n]] <-