+++ R/getVersion.R
@@ -3,5 +3,5 @@
 #' @return Current Version
 #' @export
 getVersion <- function() {
-  "0.5.00 (20181228)"
+  "0.5.01 (20181230)"
 }