+++ R/makeGroupMembers.R
@@ -15,42 +15,44 @@
 #' @param minAge integer value indicating the minimum age to consider in group
 #' formation. Pairwise kinships involving an animal of this age or younger will
 #'  be ignored. Default is 1 year.
+#' @export
 makeGroupMembers <- function(numGp, currentGroup, candidates, ped, harem,
                              minAge) {
   groupMembers <- list()
   if (harem) {
-    if (length(potentialSires(currentGroup, minAge, ped)) > 1)
+    if (length(getPotentialSires(currentGroup, minAge, ped)) > 1)
       stop(paste0("User selected to form harems with more than one male, ",
                   "There are ",
-                  length(potentialSires(currentGroup, minAge, ped)), " at ",
+                  length(getPotentialSires(currentGroup, minAge, ped)), " at ",
                   " least ", minAge, " years old in the current group."))
-    if (length(potentialSires(candidates, minAge, ped)) < numGp &
-        length(potentialSires(currentGroup, minAge, ped)) == 0)
+    if (length(currentGroup) > 0 & numGp > 1)
+      stop(paste0("User cannot have more than one breeding group with ",
+                  "animals in a current group"))
+    if (length(getPotentialSires(candidates, minAge, ped)) < numGp &
+        length(getPotentialSires(currentGroup, minAge, ped)) == 0)
       stop(paste0("User selected to form harems in ", numGp, " groups with ",
-                  "only ", length(potentialSires(currentGroup, minAge, ped)),
+                  "only ", length(getPotentialSires(currentGroup, minAge, ped)),
                   " males at least ",
                   minAge, " years old in the list of candidates."))
 
-    if (length(potentialSires(currentGroup, minAge, ped)) == 0) {
+    if (length(getPotentialSires(currentGroup, minAge, ped)) == 0) {
       ped <- ped[!is.na(ped$birth), ]
-      sires <- sample(potentialSires(candidates, minAge, ped), numGp,
+      sires <- sample(getPotentialSires(candidates, minAge, ped), numGp,
                       replace = FALSE)
       for (i in 1:numGp) {
         groupMembers[[i]] <- sires[i]
       }
     }
-  }
-  if (harem) {
-    for (i in 1:numGp) {
-      if (length(currentGroup) > 0) {
-        groupMembers[[i]] <- currentGroup
+    if (length(currentGroup) > 0) {
+      if (length(getPotentialSires(currentGroup, minAge, ped)) > 0) {
+        groupMembers[[1]] <- currentGroup
+      } else {
+        groupMembers[[1]] <- c(groupMembers[[1]], currentGroup)
       }
     }
   } else {
     for (i in 1:numGp) {
-      if (is.null(currentGroup)) { # should not occur
-        groupMembers[[i]] <- vector()
-      } else if (length(currentGroup) == 0) {
+      if (length(currentGroup) == 0) {
         groupMembers[[i]] <- vector()
       } else {
         groupMembers[[i]] <- currentGroup